tcm_qla2xxx: change tcm_qla2xxx_nacl->nport_id to u32
Fix tcm_qla2xxx_nacl->nport_id to use u32 for lookup in
to follow fc_rport->port_id in lport->lport_fcport_map.
This is currenlty used by tcm_qla2xxx_setup_nacl_from_rport()
and tcm_qla2xxx_clear_nacl_from_fcport_map() used to store
*se_node_acl.

From DanC:

The patch b3fd3cce29bb: "tcm_qla2xxx: Add >= 24xx series fabric
module for target-core" from Mar 8, 2012, leads to the following Sparse
warning:

drivers/scsi/qla2xxx/tcm_qla2xxx.c:843:37: warning: right shift by
        bigger than source value

+       domain = (nacl->nport_id >> 16) & 0xff;
                        ^^^^^^^^
nport_id is a u16 so this is always zero.

+       area = (nacl->nport_id >> 8) & 0xff;
+       al_pa = nacl->nport_id & 0xff;

[snip]

+struct tcm_qla2xxx_nacl {
+       /* From libfc struct fc_rport->port_id */
+       u16 nport_id;

fc_rport->port_id is a u32.

Reported-by: Dan Carpenter <dan.carpenter@oracle.com>

+++ drivers/scsi/qla2xxx/tcm_qla2xxx.h
@@ -10,7 +10,7 @@
 
 struct tcm_qla2xxx_nacl {
 	/* From libfc struct fc_rport->port_id */
-	u16 nport_id;
+	u32 nport_id;
 	/* Binary World Wide unique Node Name for remote FC Initiator Nport */
 	u64 nport_wwnn;
 	/* ASCII formatted WWPN for FC Initiator Nport */
